Adjust positioning and placement of .main elements
.main elements were not quite sitting the same as Google's.
In this commit, css is adjusted to help attain this.
One important adjustment is the addition of the font-size: 0px rule
to the .buttons div to remove the whitespace created in the HTML
between the two of them.
A placeholder div is added to gauge the position of the bottom of
the document, but then, as it becomes apparent that it is
floating mid-document, remove top property from .main
and add a margin instead.
A div.main-container encapsulates div.main, so that a min-width of 980px
works for the main page elements.
This commit switches the Google logo for a higher-res one and adds
an appropriate width and height element to the element.