Adjust positioning of navigation and main elements
Placement of the .main div was a little high and positioning of
navigation elements was still all shoved to the left.
With this commit, they are positioned correctly and the page is
unified as it is resized, with an appropriate min-wdith
on .navigation div.
Class names for navigation anchor elements are also changed
for better CSS usability.