Finish positioning footer and whole webpage
There are issues with html and body not working the same as google,
due to x/y overflow being different, etc.
Also, lots of padding issues and an overarching top:0 specification,
from previous inherited div traits.
With this commit, these issues are resolved and the webpage behaves
almost fully like Google.
The next issue to troubleshoot, is the fact that the footer needs to
disappear, once there is no room for it on the page like the original.