Set footer to disappear when height reduced
On Google's homepage, when reducing the height of the window, the footer
moves up with the page, until reaching the space
below the search buttons, at which point, it disappears
as the page height is decreased.
With this commit, this page imitates the behaviour of the original.
This is done by creating a container for the footer,
positioning the container absolutely and gradually sizing it via cal()
while also positioning it down, to the point where the disappear point
matches with the original.
This was not creating a full width footer, without the container also
being given a full screen width (width:100%).